Rename juce_x11 -> JuceIntegration
