--- src/core/gui/juce_x11.h
@@ -1,5 +1,0 @@
-#pragma once
-
-void juceInit();
-
-void juceIdle();
