+++ src/plugins/lv2/lv2ui.cpp
@@ -25,7 +25,7 @@
 
 #include "core/controls.h"
 #include "core/gui/MainComponent.h"
-#include "core/gui/juce_x11.h"
+#include "core/gui/JuceIntegration.h"
 #include "core/synth/PresetController.h"
 #include "core/synth/Synthesizer.h"
